Tile Map + Level Prototype
Created both a tile map and level created.

These were to test if it could be done using their systems.

Yes it can, took longer than expected but that came down to me trying to keep the code clean and not hack it.

I am not sure if it would have been worth hacking it for the speed to do it in. I think I should have in this case because outcome time should have been a bigger priority for me.